Replace in-scene overlay UI with CanvasOverlay prefab
Added generic prefab-based UI management dialog prompts (tutorialBoxManager).

Included dialogs with feedback of the wallet connection.
Play button, now prompt first to connect and authenticate before you can play.

This implementations are enough for learning, but not useful in real life. It´s neccesary to implement a server and a BD for the wallet verification and the user data storage.